% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{change_api_url}
\alias{change_api_url}
\title{Change API root URL}
\usage{
change_api_url(api_root_url)
}
\arguments{
\item{api_root_url}{Character string with the base URL of the target API.}
}
\value{
The configured API root URL.
}
\description{
Changes the base URL used by the package to connect to a compatible
statistics API. This allows switching between ISTAC, IBESTAT and any
other institution implementing the same API.
}
\examples{
change_api_url("https://datos.canarias.es/api/estadisticas/")

}
