% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_resources_classifications.R
\name{get_codelist_families_id}
\alias{get_codelist_families_id}
\title{Get codelist families}
\usage{
get_codelist_families_id(id = NULL)
}
\arguments{
\item{id}{(string) codelist family identificator}
}
\description{
This function allows to obtain a family of classifications in particular.
}
\examples{
get_codelist_families_id()
}
