% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_resources_cubes.R
\name{get_datasets_agency}
\alias{get_datasets_agency}
\title{Get datasets (agencyID)}
\usage{
get_datasets_agency(
  agencyID,
  lang = "es",
  limit = 25,
  offset = 0,
  orderBy = NULL,
  query = NULL
)
}
\arguments{
\item{agencyID}{(string) Identifier of the maintainer organization of the resource. A possible value is \code{ISTAC}.}

\item{lang}{(string) Language in which you want to get the answer.}

\item{limit}{(int) Results limit. By default \code{limit = 25}.}

\item{offset}{(int) Displacement. Result from which it is returned. By default \code{offset = 0}.}

\item{orderBy}{(string) Order. Possible values are \code{ID ASC} or \code{ID DESC},}

\item{query}{(string) Metadata query on which the searches can be built.}
}
\description{
This function allows to consult all the data sets maintained by a certain organization.
}
\examples{
get_datasets_agency(agencyID = "ISTAC")
}
