% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_auxiliary_value_tables.R
\name{get_indicators_geographic_granularities}
\alias{get_indicators_geographic_granularities}
\title{Get geographic granularities}
\usage{
get_indicators_geographic_granularities()
}
\description{
This function returns a list of geographic granularities treated
 in the ISTAC-indicators database. For example provincial, insular or
 municipal granularity.
}
\examples{
get_indicators_geographic_granularities()
}
