% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators-geographic.R
\name{get_indicators_geographical_values}
\alias{get_indicators_geographical_values}
\title{Get geographical values}
\usage{
get_indicators_geographical_values(
  geographicalgranularitycode,
  subjectcode,
  systemcode = ""
)
}
\arguments{
\item{geographicalgranularitycode}{(string): geographical granularity code}

\item{subjectcode}{(string): subject code}

\item{systemcode}{(string): system code}
}
\description{
This function returns values of a geographical granularity that in turn are part of a
specific theme or system of indicators.
}
\examples{
get_indicators_geographical_values(
"REGIONS",
subjectcode="051",
systemcode="C00067A"
)
}
