% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_auxiliary_value_tables.R
\name{get_indicators_subjects}
\alias{get_indicators_subjects}
\title{Get subjects}
\usage{
get_indicators_subjects()
}
\description{
This function returns all subjects which the ISTAC classifies
 its statistical operations.
}
\examples{
get_indicators_subjects()
}
