% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_indicators_systems.R
\name{get_indicators_systems_code}
\alias{get_indicators_systems_code}
\title{Get indicators system code}
\usage{
get_indicators_systems_code(indicatorSystemCode)
}
\arguments{
\item{indicatorSystemCode}{(string) an indicator system code}
}
\description{
This function returns metadata of a system of indicators
 published in the ISTAC-indicators database. The indicators are simple or
 compound statistics, however a single indicator can rarely provide useful
 information about complex phenomena such as the economic situation, living
 conditions, schooling or others.
}
\examples{
get_indicators_systems_code("C00075H")
}
