% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators-systems.R
\name{get_indicators_systems_code_instances_code}
\alias{get_indicators_systems_code_instances_code}
\title{Get indicators system code instances code}
\usage{
get_indicators_systems_code_instances_code(
  indicatorsystemcode,
  indicatorinstancecode
)
}
\arguments{
\item{indicatorsystemcode}{(string) indicator system code}

\item{indicatorinstancecode}{(string) indicator instance code}
}
\description{
This function returns metadata of an indicator set associated with a specific indicator
system. An instance of an indicator is nothing more than a spatio-temporal query of an
indicator when it is incorporated into a specific indicator system.
}
\examples{
get_indicators_systems_code_instances_code(
"C00075H",
"21af0477-d63b-493b-ad02-4ab181547223"
)
}
