% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_auxiliary_value_tables.R
\name{get_indicators_time_granularities}
\alias{get_indicators_time_granularities}
\title{Get time granularities}
\usage{
get_indicators_time_granularities()
}
\description{
This function returns a list of temporary granularity treated
 in the ISTAC data bank-indicators ordered from highest to lowest granularity.
 For example annual, quarterly or monthly granularity.
}
\examples{
get_indicators_time_granularities()
}
