% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_resources_category.R
\name{get_schemes_agency_resource_version}
\alias{get_schemes_agency_resource_version}
\title{Get category schemes agency resource version}
\usage{
get_schemes_agency_resource_version(agencyID, resourceID, version)
}
\arguments{
\item{agencyID}{(string) Identifier of the agency that publishes.}

\item{resourceID}{(string) Resource identifier.}

\item{version}{(string) Specific version of the resource.}
}
\description{
This function returns the content from /v1.0/categoryschemes/{agencyID}/{resourceID}/{version}
}
\examples{
get_schemes_agency_resource_version("ISTAC", "TEMAS_CANARIAS", "01.000")
}
