% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_resources_category.R
\name{get_schemes_agency_resource_version_categories_id}
\alias{get_schemes_agency_resource_version_categories_id}
\title{Get category schemes agency resource version categories (id)}
\usage{
get_schemes_agency_resource_version_categories_id(agencyID, resourceID,
  version, categoryID)
}
\arguments{
\item{agencyID}{(string) Identifier of the agency that publishes.}

\item{resourceID}{(string) Resource identifier.}

\item{version}{(string) Specific version of the resource.}

\item{categoryID}{(string)}
}
\description{
This function returns the content from /v1.0/categoryschemes/{agencyID}/{resourceID}/{version}/categories/{categoryID}
}
\examples{
get_schemes_agency_resource_version_categories_id("ISTAC", "TEMAS_CANARIAS", "01.000", "060")
}
