% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical-resources-cubes.R
\name{get_statisticalresources_datasets_agency}
\alias{get_statisticalresources_datasets_agency}
\title{Get datasets (agencyID)}
\usage{
get_statisticalresources_datasets_agency(
  agencyid,
  lang = "es",
  limit = 25,
  offset = 0,
  orderby = "",
  query = ""
)
}
\arguments{
\item{agencyid}{(string): Identifier of the maintainer organization of the resource. A possible value is \code{ISTAC}.}

\item{lang}{(string): Language in which you want to get the answer.}

\item{limit}{(int): Results limit. By default \code{limit=25}.}

\item{offset}{(int): Displacement. Result from which it is returned. By default \code{offset=0}.}

\item{orderby}{(string): Order. Possible values are \verb{ID ASC} or \verb{ID DESC}.}

\item{query}{(string): Metadata query on which the searches can be built.}
}
\description{
This function allows to consult all the data sets maintained by a certain organization.
}
\examples{
get_statisticalresources_datasets_agency(agencyid="ISTAC")
}
