% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-resources-category.R
\name{get_structuralresources_categorisations_agency}
\alias{get_structuralresources_categorisations_agency}
\title{Get categorisations agency}
\usage{
get_structuralresources_categorisations_agency(
  agencyid,
  limit = 25,
  offset = 0,
  query = "",
  orderby = ""
)
}
\arguments{
\item{agencyid}{(string): Identifier of the agency that publishes.}

\item{limit}{(int): Results limit. By default \code{limit = 25}.}

\item{offset}{(int): Displacement. Result from which it is returned. By default \code{offset=0}.}

\item{query}{(string): Query to filter the results.}

\item{orderby}{(string): Field by which to sort the results.}
}
\description{
This function returns the content from \verb{/v1.0/categorisations/\{agencyID\}}
}
\examples{
get_structuralresources_categorisations_agency("ISTAC")
}
