% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-resources-category.R
\name{get_structuralresources_category_schemes}
\alias{get_structuralresources_category_schemes}
\title{Get category schemes}
\usage{
get_structuralresources_category_schemes(
  limit = 25,
  offset = 0,
  query = "",
  orderby = ""
)
}
\arguments{
\item{limit}{(int): Results limit. By default \code{limit = 25}.}

\item{offset}{(int): Displacement. Result from which it is returned.  By default \code{offset = 0}.}

\item{query}{(string): Query to filter the results.}

\item{orderby}{(string): Field by which to sort the results.}
}
\description{
This function returns the content from \verb{/v1.0/categoryschemes}
}
\examples{
get_structuralresources_category_schemes()
get_structuralresources_category_schemes(
query="ID EQ 2090",
orderby="ID ASC"
)
}
