% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-resources-category.R
\name{get_structuralresources_category_schemes_agency_resource_version_categories_id}
\alias{get_structuralresources_category_schemes_agency_resource_version_categories_id}
\title{Get category schemes agency resource version categories (id)}
\usage{
get_structuralresources_category_schemes_agency_resource_version_categories_id(
  agencyid,
  resourceid,
  version,
  categoryid
)
}
\arguments{
\item{agencyid}{(string): Identifier of the agency that publishes.}

\item{resourceid}{(string): Resource identifier.}

\item{version}{(string): Specific version of the resource.}

\item{categoryid}{(string): category identifier.}
}
\description{
This function returns the content from \verb{/v1.0/categoryschemes/\{agencyID\}/\{resourceID\}/\{version\}/categories/\{categoryID\}}
}
\examples{
get_structuralresources_category_schemes_agency_resource_version_categories_id(
"ISTAC",
"TEMAS_CANARIAS",
"01.000",
"060"
)
get_structuralresources_category_schemes_agency_resource_version_categories_id(
"ISTAC",
"TEMAS_CANARIAS",
"01.000",
"060.060_010.060_010_010"
)
}
