% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-resources-clasifications.R
\name{get_structuralresources_codelist_families}
\alias{get_structuralresources_codelist_families}
\title{Get codelist families}
\usage{
get_structuralresources_codelist_families(
  limit = 25,
  offset = 0,
  orderby = "",
  query = ""
)
}
\arguments{
\item{limit}{(int): Results limit. By default \code{limit = 25}.}

\item{offset}{(int): Displacement. Result from which it is returned. By default \code{offset = 0}.}

\item{orderby}{(string): Field by which to sort the results.}

\item{query}{(string): Query to filter the results.}
}
\description{
This function returns the list of families of classifications
}
\examples{
get_structuralresources_codelist_families()
}
