% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-resources-clasifications.R
\name{get_structuralresources_codelist_families_id}
\alias{get_structuralresources_codelist_families_id}
\title{Get codelist families}
\usage{
get_structuralresources_codelist_families_id(id)
}
\arguments{
\item{id}{(string): codelist family identificator}
}
\description{
This function allows to obtain a family of classifications in particular.
}
\examples{
get_structuralresources_codelist_families_id('CODELIST_ID')
}
