% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-resources-clasifications.R
\name{get_structuralresources_codelists}
\alias{get_structuralresources_codelists}
\title{Get codelists}
\usage{
get_structuralresources_codelists(
  limit = 25,
  offset = 0,
  query = "",
  orderby = ""
)
}
\arguments{
\item{limit}{(int): Results limit. By default \code{limit = 25}.}

\item{offset}{(int): Displacement. Result from which it is returned. By default \code{offset = 0}.}

\item{query}{(string): Query to filter the results.}

\item{orderby}{(string): Field by which to sort the results.}
}
\description{
This function allows to obtain the list of classifications.
}
\examples{
get_structuralresources_codelists()
}
