% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-resources-clasifications.R
\name{get_structuralresources_codelists_agency_resource}
\alias{get_structuralresources_codelists_agency_resource}
\title{Get codelists agency resource}
\usage{
get_structuralresources_codelists_agency_resource(
  agencyid,
  resourceid,
  limit = 25,
  offset = 0,
  query = "",
  orderby = ""
)
}
\arguments{
\item{agencyid}{(string): Agency identificator.}

\item{resourceid}{(string): Resource identificator.}

\item{limit}{(int): Results limit. By default \code{limit = 25}.}

\item{offset}{(int): Displacement. Result from which it is returned. By default \code{offset = 0}.}

\item{query}{(string): Query to filter the results.}

\item{orderby}{(string): Field by which to sort the results.}
}
\description{
This function allows to obtain all the versions of a classification with a certain
identifier and that is also kept by a certain organization.
}
\examples{
get_structuralresources_codelists_agency_resource("ISTAC", "CL_AREA_ES")
}
