% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_resources_classifications.R
\name{get_structuralresources_codelists_agency_resource_version}
\alias{get_structuralresources_codelists_agency_resource_version}
\title{Get codelists agency resource version}
\usage{
get_structuralresources_codelists_agency_resource_version(
  agencyID,
  resourceID,
  version
)
}
\arguments{
\item{agencyID}{(string) Agency identificator.}

\item{resourceID}{(string) Resource identificator.}

\item{version}{(string) Specific resource version.}
}
\description{
This function allows you to consult a particular version of a classification.
}
\examples{
get_structuralresources_codelists_agency_resource_version("ISTAC", "CL_AREA_ES", "01.000")
}
