% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-resources-clasifications.R
\name{get_structuralresources_codelists_agency_resource_version_codes_codeid}
\alias{get_structuralresources_codelists_agency_resource_version_codes_codeid}
\title{Get codelists agency resource version codes (codeID)}
\usage{
get_structuralresources_codelists_agency_resource_version_codes_codeid(
  agencyid,
  resourceid,
  version,
  codeid
)
}
\arguments{
\item{agencyid}{(string): Agency identificator.}

\item{resourceid}{(string): Resource identificator.}

\item{version}{(string): Specific resource version.}

\item{codeid}{(string): Code identificator.}
}
\description{
This function allows to consult a specific code of a version of a classification.
}
\examples{
get_structuralresources_codelists_agency_resource_version_codes_codeid(
"ISTAC", "CL_AREA_ES", "01.000", "ES706A01")
}
