% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-resources-clasifications.R
\name{get_structuralresources_codelists_agency_resource_version_restrictions}
\alias{get_structuralresources_codelists_agency_resource_version_restrictions}
\title{Get codelists agency resource version restrictions}
\usage{
get_structuralresources_codelists_agency_resource_version_restrictions(
  agencyid,
  resourceid,
  version
)
}
\arguments{
\item{agencyid}{(string): Agency identificator.}

\item{resourceid}{(string): Resource identificator.}

\item{version}{(string): Specific resource version.}
}
\description{
This function allows you to consult any restrictions for a specific code of a version of a classification.
}
\examples{
get_structuralresources_codelists_agency_resource_version_restrictions(
"ISTAC",
"CL_AREA_ES",
"~latest"
)
}
