% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-resources-concepts.R
\name{get_structuralresources_concept_schemes_agency_resource}
\alias{get_structuralresources_concept_schemes_agency_resource}
\title{Get concept schemes agency resource}
\usage{
get_structuralresources_concept_schemes_agency_resource(
  agencyid,
  resourceid,
  limit = 25,
  offset = 0,
  query = "",
  orderby = ""
)
}
\arguments{
\item{agencyid}{(string): Identifier of the agency that publishes.}

\item{resourceid}{(string): Resource identifier.}

\item{limit}{(int): Results limit. By default \code{limit = 25}.}

\item{offset}{(int): Displacement. Result from which it is returned. By default \code{offset = 0}.}

\item{query}{(string): Query to filter the results.}

\item{orderby}{(string): Field by which to sort the results.}
}
\description{
This function returns the content from \verb{/v1.0/conceptschemes/\{agencyID\}/\{resourceID\}}
}
\examples{
get_structuralresources_concept_schemes_agency_resource(
agencyid="ISTAC",
resourceid="CSM_C00010A_SIE"
)
}
