% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-resources-concepts.R
\name{get_structuralresources_concept_types}
\alias{get_structuralresources_concept_types}
\title{Get concept types}
\usage{
get_structuralresources_concept_types()
}
\description{
This function returns the content from \verb{/v1.0/conceptTypes}
}
\examples{
get_structuralresources_concept_types()
}
