% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-resources-datastructures.R
\name{get_structuralresources_content_constraints_agency_resource_version}
\alias{get_structuralresources_content_constraints_agency_resource_version}
\title{Get content constraints agency resource version}
\usage{
get_structuralresources_content_constraints_agency_resource_version(
  agencyid,
  resourceid,
  version
)
}
\arguments{
\item{agencyid}{(string): Identifier of the agency that publishes.}

\item{resourceid}{(string): Resource identifier.}

\item{version}{(string): Specific version of the resource.}
}
\description{
This function returns the content from \verb{/v1.0/contentConstraints/\{agencyID\}/\{resourceID\}/\{version\}}
}
\examples{
get_structuralresources_content_constraints_agency_resource_version(
"ISTAC",
"CSM_C00010A_SIE",
"01.000"
)
}
