% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-resources-datastructures.R
\name{get_structuralresources_content_constraints_agency_resource_version_regions}
\alias{get_structuralresources_content_constraints_agency_resource_version_regions}
\title{Get content constraints agency resource version regions}
\usage{
get_structuralresources_content_constraints_agency_resource_version_regions(
  regioncode,
  agencyid,
  resourceid,
  version
)
}
\arguments{
\item{regioncode}{(string) Region code.}

\item{agencyid}{(string) Identifier of the agency that publishes.}

\item{resourceid}{(string) Resource identifier.}

\item{version}{(string) Specific version of the resource.}
}
\description{
This function returns the content from
\verb{/v1.0/contentConstraints/\{agencyID\}/\{resourceID\}/\{version\}/regions/\{regionCode\}}
}
\examples{
get_structuralresources_content_constraints_agency_resource_version_regions(
"0001",
"ISTAC",
"CSM_C00010A_SIE",
"01.000"
)
}
