% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-resources-variables.R
\name{get_structuralresources_geoinfo}
\alias{get_structuralresources_geoinfo}
\title{Get geoinfo}
\usage{
get_structuralresources_geoinfo(
  variableid,
  resourceid,
  fields = "",
  limit = 25,
  offset = 0,
  query = "",
  orderby = ""
)
}
\arguments{
\item{variableid}{(string): Variable identificator.}

\item{resourceid}{(string): Resource identificator.}

\item{fields}{(string): Additional fields that you want to show in the answer.}

\item{limit}{(int): Results limit. By default \code{limit = 25}.}

\item{offset}{(int): Displacement. Result from which it is returned. By default \code{offset = 0}.}

\item{query}{(string): Query to filter the results.}

\item{orderby}{(string): Field by which to sort the results.}
}
\description{
This function returns data from
\verb{/v1.0/variables/\{variableID\}/variableelements/\{resourceID\}/geoinfo}
}
\examples{
get_structuralresources_geoinfo("VR_TERRITORIO", "MUN_ICOD_VINOS")
}
