% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-resources-variables.R
\name{get_structuralresources_variable_families_id}
\alias{get_structuralresources_variable_families_id}
\title{Get variable families (id)}
\usage{
get_structuralresources_variable_families_id(id)
}
\arguments{
\item{id}{(string): Variable family identificator.}
}
\description{
This function returns data from \verb{/v1.0/variablefamilies/\{id\}}
}
\examples{
get_structuralresources_variable_families_id("VRF_DEMOGRAFICAS")
}
