% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-resources-variables.R
\name{get_structuralresources_variables_id}
\alias{get_structuralresources_variables_id}
\title{Get variables (id)}
\usage{
get_structuralresources_variables_id(id)
}
\arguments{
\item{id}{(string): Variable identifier.}
}
\description{
This function returns data from \verb{/v1.0/variables/\{id\}}
}
\examples{
get_structuralresources_variables_id("VR_SEXO")
}
