% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_resources_variables.R
\name{get_variables_id}
\alias{get_variables_id}
\title{Get variables (id)}
\usage{
get_variables_id(id)
}
\arguments{
\item{id}{(string) Variable identificator.}
}
\description{
This function returns data from /v1.0/variables/{id}
}
\examples{
get_variables_id("VR_SEXO")
}
