% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_indicators.R
\name{get_indicators_code}
\alias{get_indicators_code}
\title{Get indicators code}
\usage{
get_indicators_code(indicatorCode)
}
\arguments{
\item{indicatorCode}{(string) an indicator code}
}
\description{
This function returns the metadata that describe the characteristics
 of a specific indicator, allowing the compression of the measured fact;
 also through the data request the complete data (for all spacetime) of the
 indicator is provided.
}
\examples{
get_indicators_code("AFILIACIONES")
get_indicators_code("PARO_REGISTRADO")
}
