% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_auxiliary_value_tables.R
\name{get_indicators_geographical_values}
\alias{get_indicators_geographical_values}
\title{Get geographical values}
\usage{
get_indicators_geographical_values(
  geographicalGranularityCode,
  subjectCode = NULL,
  systemCode = NULL
)
}
\arguments{
\item{geographicalGranularityCode}{(string) geographical granularity code}

\item{subjectCode}{(string) subject code}

\item{systemCode}{(string) system code}
}
\description{
This function returns values of a geographical granularity that
 in turn are part of a specific theme or system of indicators.
}
\examples{
get_indicators_geographical_values("REGIONS")
get_indicators_geographical_values("REGIONS", subjectCode = "051", systemCode = "C00067A")
}
