% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_indicators_systems.R
\name{get_indicators_systems}
\alias{get_indicators_systems}
\title{Get indicators systems}
\usage{
get_indicators_systems(limit = 25, offset = 0)
}
\arguments{
\item{limit}{(int) Results limit. By default \code{limit = 25}.}

\item{offset}{(int) Displacement. Result from which it is returned. By default \code{offset = 0}.}
}
\description{
This function returns a list of indicator systems published in
 the ISTAC-indicators database. The indicators are simple or compound
 statistics, however a single indicator can rarely provide useful
 information about complex phenomena such as the economic situation,
 living conditions, schooling or others. Indicator systems are generally
 designed to generate more and more accurate information about the
 conditions of a phenomenon; and for this they are organized in dimensions
 or areas of analysis, under which the indicators are integrated.
}
\examples{
get_indicators_systems()
}
