% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators-systems.R
\name{get_indicators_systems_code_instances_code_data}
\alias{get_indicators_systems_code_instances_code_data}
\title{Get indicators system code instances code data}
\usage{
get_indicators_systems_code_instances_code_data(
  indicatorsystemcode,
  indicatorinstancecode,
  representation = "",
  granularity = "",
  fields = ""
)
}
\arguments{
\item{indicatorsystemcode}{(string): Indicator system code}

\item{indicatorinstancecode}{(string): Indicator instance code}

\item{representation}{(string): Allows filtering the observations by their value.}

\item{granularity}{(string): Allows to filter the observations through the granularities of the same.}

\item{fields}{(string): Allows you to customize the response by excluding fields. The possible values are: ``-observationsMetadata``.}
}
\description{
This function returns metadata of an indicator set associated with a specific indicator
system. An instance of an indicator is nothing more than a spatio-temporal query of an
indicator when it is incorporated into a specific indicator system.
}
\examples{
get_indicators_systems_code_instances_code_data(
"C00075H",
"21af0477-d63b-493b-ad02-4ab181547223"
)
}
