% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical-resources-cubes.R
\name{get_statisticalresources_datasets}
\alias{get_statisticalresources_datasets}
\title{Get datasets}
\usage{
get_statisticalresources_datasets(
  lang = "es",
  limit = 25,
  offset = 0,
  orderby = "",
  query = ""
)
}
\arguments{
\item{lang}{(string): Language in which you want to get the answer.}

\item{limit}{(int): Results limit. By default \code{limit=25}.}

\item{offset}{(int): Displacement. Result from which it is returned. By default \code{offset=0}.}

\item{orderby}{(string): Order. Possible values are \verb{ID ASC} or \verb{ID DESC}.}

\item{query}{(string): Metadata query on which the searches can be built.}
}
\description{
This function allows consulting all existing statistical data cubes.
}
\examples{
get_statisticalresources_datasets(limit=5)
}
