% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical-resources-cubes.R
\name{get_statisticalresources_datasets_agency_resource}
\alias{get_statisticalresources_datasets_agency_resource}
\title{Get datasets (agencyID / resourceID)}
\usage{
get_statisticalresources_datasets_agency_resource(
  agencyid,
  resourceid,
  lang = "es",
  limit = 25,
  offset = 0,
  orderby = "",
  query = ""
)
}
\arguments{
\item{agencyid}{(string): Identifier of the maintainer organization of the resource. A possible value is ``ISTAC``.}

\item{resourceid}{(string): Resource identifier. A possible value is ``C00010A_000002``.}

\item{lang}{(string): Language in which you want to get the answer.}

\item{limit}{(int): Results limit. By default ``limit=25``.}

\item{offset}{(int): Displacement. Result from which it is returned. By default ``offset=0``.}

\item{orderby}{(string): Order. Possible values are ``ID ASC`` or ``ID DESC``.}

\item{query}{(string): Metadata query on which the searches can be built.}
}
\description{
This function allows to obtain all the versions of a statistical cube with a certain
identifier and that also maintains a certain organization.
}
\examples{
get_statisticalresources_datasets_agency_resource(
agencyid="ISTAC",
resourceid="C00010A_000002"
)
}
