% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical-resources-queries.R
\name{get_statisticalresources_queries}
\alias{get_statisticalresources_queries}
\title{Get queries}
\usage{
get_statisticalresources_queries(
  lang = "es",
  limit = 25,
  offset = 0,
  orderby = "",
  query = ""
)
}
\arguments{
\item{lang}{(string): Language in which you want to get the answer.}

\item{limit}{(int): Results limit. By default \code{limit=25}.}

\item{offset}{(int): Displacement. Result from which it is returned. By default \code{offset=0}.}

\item{orderby}{(string): Order. Possible values are \verb{ID ASC} or \verb{ID DESC},}

\item{query}{(string): Metadata query on which the searches can be built.}
}
\description{
This function allows consulting all existing statistical queries.
}
\examples{
get_statisticalresources_queries(limit=5)
}
