% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical-resources-queries.R
\name{get_statisticalresources_queries_agency_resource}
\alias{get_statisticalresources_queries_agency_resource}
\title{Get queries (agencyID / resourceID)}
\usage{
get_statisticalresources_queries_agency_resource(
  agencyid,
  resourceid,
  fields = "",
  lang = "es",
  as_dataframe = FALSE
)
}
\arguments{
\item{agencyid}{(string): Identifier of the maintainer organization of the resource. A possible value is ``ISTAC``.}

\item{resourceid}{(string): Resource identifier. A possible value is ``C00010A_000002``.}

\item{fields}{(string): Allows you to customize the response by excluding fields from it.
- The possible values are ``-metadata`` and ``-data``.}

\item{lang}{(string): Language in which you want to get the answer.}

\item{as_dataframe}{(bool): If True, this function returns a namedtuple with:
- dataframe: pandas dataframe built from API response.
- codelists: mapping between codes and representations for each column.}
}
\description{
This function allows to obtain final data of a statistical query with a certain identifier and that also maintains a certain organization.
}
\examples{
get_statisticalresources_queries_agency_resource(
agencyid="ISTAC",
resourceid="C00017A_000001"
)
}
