% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_resources_category.R
\name{get_structuralresources_categorisations_agency_resource_version}
\alias{get_structuralresources_categorisations_agency_resource_version}
\title{Get categorisations agency resource version}
\usage{
get_structuralresources_categorisations_agency_resource_version(
  agencyID,
  resourceID,
  version
)
}
\arguments{
\item{agencyID}{(string) Identifier of the agency that publishes.}

\item{resourceID}{(string) Resource identifier.}

\item{version}{(string) Specific version of the resource.}
}
\description{
This function returns the content from /v1.0/categorisations/{agencyID}/{resourceID}/{version}
}
\examples{
get_structuralresources_categorisations_agency_resource_version("ISTAC", "cat2", "01.000")
}
