% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-resources-category.R
\name{get_structuralresources_category_schemes_agency_resource_version_categories}
\alias{get_structuralresources_category_schemes_agency_resource_version_categories}
\title{Get category schemes agency resource version categories}
\usage{
get_structuralresources_category_schemes_agency_resource_version_categories(
  agencyid,
  resourceid,
  version,
  limit = 25,
  offset = 0,
  query = "",
  orderby = ""
)
}
\arguments{
\item{agencyid}{(string): Identifier of the agency that publishes.}

\item{resourceid}{(string): Resource identifier.}

\item{version}{(string): Specific version of the resource.}

\item{limit}{(int): Results limit. By default ``limit = 25``.}

\item{offset}{(int): Displacement. Result from which it is returned. By default ``offset = 0``.}

\item{query}{(string): Query to filter the results.}

\item{orderby}{(string): Field by which to sort the results.}
}
\description{
This function returns the content from ``/v1.0/categoryschemes/{agencyID}/{resourceID}/{version}/categories``
}
\examples{
get_structuralresources_category_schemes_agency_resource_version_categories(
"ISTAC",
"TEMAS_CANARIAS",
"01.000"
)
}
