% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_resources_classifications.R
\name{get_structuralresources_codelists_agency}
\alias{get_structuralresources_codelists_agency}
\title{Get codelists agency}
\usage{
get_structuralresources_codelists_agency(
  agencyID,
  limit = 25,
  offset = 0,
  query = NULL,
  orderBy = NULL
)
}
\arguments{
\item{agencyID}{(string) Agency identificator.}

\item{limit}{(int) Results limit. By default \code{limit = 25}.}

\item{offset}{(int) Displacement. Result from which it is returned.  By default \code{offset = 0}.}

\item{query}{(string) Query to filter the results.}

\item{orderBy}{(string) Field by which to sort the results.}
}
\description{
This function allows obtaining the list of all the classifications
 maintained by a certain organization.
}
\examples{
get_structuralresources_codelists_agency("ISTAC")
get_structuralresources_codelists_agency("ESTAT")
}
