% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-resources-clasifications.R
\name{get_structuralresources_codelists_agency_resource_version_codes}
\alias{get_structuralresources_codelists_agency_resource_version_codes}
\title{Get codelists agency resource version codes}
\usage{
get_structuralresources_codelists_agency_resource_version_codes(
  agencyid,
  resourceid,
  version,
  limit = 25,
  offset = 0,
  query = "",
  orderby = "",
  openness = "",
  order = "",
  fields = ""
)
}
\arguments{
\item{agencyid}{(string): Agency identificator.}

\item{resourceid}{(string): Resource identificator.}

\item{version}{(string): Specific resource version.}

\item{limit}{(int): Results limit. By default ``limit = 25``.}

\item{offset}{(int): Displacement. Result from which it is returned. By default ``offset = 0``.}

\item{query}{(string): Query to filter the results. orderby (string): Field by which to sort the results.}

\item{orderby}{(string): Order established for data.}

\item{openness}{(string): Opening established for viewing.}

\item{order}{(string): Order established for visualization.}

\item{fields}{(string): Additional fields that you want to show in the answer.}
}
\description{
This function allows to consult the codes of a version of a classification. Note that
if wildcards are used as ``~all`` or one of the ``limit``, ``offset``, ``query`` or
``orderBy`` parameters, the list will be automatically paginated.
}
\examples{
get_structuralresources_codelists_agency_resource_version_codes(
"ISTAC",
"CL_AREA_ES",
"01.000"
)
}
