% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-resources-concepts.R
\name{get_structuralresources_concept_schemes_agency_resource_version_concepts_id}
\alias{get_structuralresources_concept_schemes_agency_resource_version_concepts_id}
\title{Get concept schemes agency resource version concepts (id)}
\usage{
get_structuralresources_concept_schemes_agency_resource_version_concepts_id(
  agencyid,
  resourceid,
  version,
  conceptid
)
}
\arguments{
\item{agencyid}{(string): Identifier of the agency that publishes.}

\item{resourceid}{(string): Resource identifier.}

\item{version}{(string): Specific version of the resource.}

\item{conceptid}{(string): Concept identifier.}
}
\description{
This function returns the content from
``/v1.0/conceptschemes/{agencyID}/{resourceID}/{version}/concepts/{conceptID}``
}
\examples{
get_structuralresources_concept_schemes_agency_resource_version_concepts_id(
agencyid="ISTAC",
resourceid="CSM_C00010A_SIE",
version="01.000",
conceptid="ELECTORES"
)
}
