% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_resources_concepts.R
\name{get_structuralresources_concept_schemes_agency_resource_version_concepts_id}
\alias{get_structuralresources_concept_schemes_agency_resource_version_concepts_id}
\title{Get concept schemes agency resource version concepts (id)}
\usage{
get_structuralresources_concept_schemes_agency_resource_version_concepts_id(
  agencyID,
  resourceID,
  version,
  conceptID
)
}
\arguments{
\item{agencyID}{(string) Identifier of the agency that publishes.}

\item{resourceID}{(string) Resource identifier.}

\item{version}{(string) Specific version of the resource.}

\item{conceptID}{(string) Concept identifier.}
}
\description{
This function returns the content from /v1.0/conceptschemes/{agencyID}/{resourceID}/{version}/concepts/{conceptID}
}
\examples{
get_structuralresources_concept_schemes_agency_resource_version_concepts_id(
   agencyID = "ISTAC", resourceID = "CL_AREA_ES", version = "01.000", conceptID = 0
)
}
