% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-resources-datastructures.R
\name{get_structuralresources_content_constraints_agency}
\alias{get_structuralresources_content_constraints_agency}
\title{Get content constraints agency}
\usage{
get_structuralresources_content_constraints_agency(
  agencyid,
  limit = 25,
  offset = 0,
  query = "",
  orderby = ""
)
}
\arguments{
\item{agencyid}{(string): Identifier of the agency that publishes.}

\item{limit}{(int): Results limit. By default ``limit = 25``.}

\item{offset}{(int): Displacement. Result from which it is returned. By default ``offset = 0``.}

\item{query}{(string): Query to filter the results.}

\item{orderby}{(string): Field by which to sort the results.}
}
\description{
This function returns the content from ``/v1.0/contentConstraints/{agencyID}``
}
\examples{
get_structuralresources_content_constraints_agency("ISTAC")
}
