% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_resources_datastructures.R
\name{get_structuralresources_content_constraints_agency_resource_version_regions}
\alias{get_structuralresources_content_constraints_agency_resource_version_regions}
\title{Get content constraints agency resource version regions}
\usage{
get_structuralresources_content_constraints_agency_resource_version_regions(
  regionCode,
  agencyID,
  resourceID,
  version
)
}
\arguments{
\item{regionCode}{(string) Region code.}

\item{agencyID}{(string) Identifier of the agency that publishes.}

\item{resourceID}{(string) Resource identifier.}

\item{version}{(string) Specific version of the resource.}
}
\description{
This function returns the content from /v1.0/contentConstraints/{agencyID}/{resourceID}/{version}/regions/{regionCode}
}
\examples{
get_structuralresources_content_constraints_agency_resource_version_regions(
   "0001", "ISTAC", "CL_AREA", "01.000"
)
}
