% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-resources-datastructures.R
\name{get_structuralresources_data_structures}
\alias{get_structuralresources_data_structures}
\title{Get data structures}
\usage{
get_structuralresources_data_structures(
  limit = 25,
  offset = 0,
  query = "",
  orderby = ""
)
}
\arguments{
\item{limit}{(int): Results limit. By default ``limit = 25``.}

\item{offset}{(int): Displacement. Result from which it is returned. By default ``offset = 0``.}

\item{query}{(string): Query to filter the results.}

\item{orderby}{(string): Field by which to sort the results.}
}
\description{
This function returns the content from ``/v1.0/datastructures``
}
\examples{
get_structuralresources_data_structures()
get_structuralresources_data_structures(
query="ID EQ 2090",
orderby="ID ASC"
)
}
