% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-resources-datastructures.R
\name{get_structuralresources_data_structures_agency_resource_version}
\alias{get_structuralresources_data_structures_agency_resource_version}
\title{Get data structures agency resource version}
\usage{
get_structuralresources_data_structures_agency_resource_version(
  agencyid,
  resourceid,
  version
)
}
\arguments{
\item{agencyid}{(string): Identifier of the agency that publishes.}

\item{resourceid}{(string): Resource identifier.}

\item{version}{(string): Specific version of the resource.}
}
\description{
This function returns the content from
``/v1.0/datastructures/{agencyID}/{resourceID}/{version}``
}
\examples{
get_structuralresources_data_structures_agency_resource_version(
"ISTAC",
"DSD_C00010A_00001",
"01.001"
)
}
