% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_resources_variables.R
\name{get_structuralresources_geoinfo}
\alias{get_structuralresources_geoinfo}
\title{Get geoinfo}
\usage{
get_structuralresources_geoinfo(
  variableID,
  resourceID,
  fields = NULL,
  limit = 25,
  offset = 0,
  query = NULL,
  orderBy = NULL
)
}
\arguments{
\item{variableID}{(string) Variable identificator.}

\item{resourceID}{(string) Resource identificator.}

\item{fields}{(string) Additional fields that you want to show in the answer.}

\item{limit}{(int) Results limit. By default \code{limit = 25}.}

\item{offset}{(int) Displacement. Result from which it is returned.  By default \code{offset = 0}.}

\item{query}{(string) Query to filter the results.}

\item{orderBy}{(string) Field by which to sort the results.}
}
\description{
This function returns data from /v1.0/variables/{variableID}/variableelements/{resourceID}/geoinfo
}
\examples{
geoinfo <- get_structuralresources_geoinfo("VR_TERRITORIO", "MUN_ICOD_VINOS")
}
