% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_resources_variables.R
\name{get_structuralresources_variable_families_id_variables}
\alias{get_structuralresources_variable_families_id_variables}
\title{Get variable families (id) variables}
\usage{
get_structuralresources_variable_families_id_variables(
  id,
  limit = 25,
  offset = 0,
  query = NULL,
  orderBy = NULL
)
}
\arguments{
\item{id}{(string) Variable family identificator.}

\item{limit}{(int) Results limit. By default \code{limit = 25}.}

\item{offset}{(int) Displacement. Result from which it is returned.  By default \code{offset = 0}.}

\item{query}{(string) Query to filter the results.}

\item{orderBy}{(string) Field by which to sort the results.}
}
\description{
This function returns data from /v1.0/variablefamilies/{id}/variables
}
\examples{
get_structuralresources_variable_families_id_variables("VRF_DEMOGRAFICAS")
}
