% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-resources-variables.R
\name{get_structuralresources_variableelements_resource}
\alias{get_structuralresources_variableelements_resource}
\title{Get variableelements resource}
\usage{
get_structuralresources_variableelements_resource(variableid, resourceid)
}
\arguments{
\item{variableid}{(string): Resource identificator.}

\item{resourceid}{(string): Variable identificator.}
}
\description{
This function returns data from ``/v1.0/variables/{variableID}/variableelements/{resourceID}``
}
\examples{
get_structuralresources_variableelements_resource("VR_SEXO", "FEMALE")
}
